Default cause now displays all npc names.
The default case now displays all npc names by looping through the
definitions. This is useful because you don't have to manually add it
yourself. An extra parameter was added because the first npcType gets
reset to -1 once a player clicks on the dialogue option causing an
exception. This is a cheap hack right now, will fix later.