Using complete 602 ItemDefinitions
I did this so whoever uses this base are able to use items all the way
up to 602 revision. This gives whoever uses this flexibility in the kind
of server they want to create. I did not go with old school items
because majority of the items changed and I feel this will limit this
base to only old school revisions. However this can easily be changed to
make an old school revision.